pyproject.toml
[tool.poetry]
name = "uk-bike-usage-analysis"
version = "0.1.0"
description = "An end-to-end DE project for UK bike data"
authors = ["Mikaelnystroms <mikaelnystromm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
gcloud = "^0.18.3"
google-cloud = "^0.34.0"
google-cloud-storage = "^2.16.0"
duckdb = "^0.10.1"

[tool.poetry.group.dev.dependencies]
prefect = "^2.16.9"
requests = "^2.31.0"
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
